chore: remove market data requirement from swappers
