+++ src/lib/swapper/swapper.ts
@@ -5,14 +5,13 @@ import { isFulfilled as isFulfilledPredicate, timeout } from 'lib/utils'
 
 import type { GetTradeQuoteInput, SwapErrorRight, SwapperName, TradeQuote2 } from './api'
 import { QUOTE_TIMEOUT_ERROR, QUOTE_TIMEOUT_MS, swappers } from './constants'
-import type { QuoteResult, TradeQuoteDeps } from './types'
+import type { QuoteResult } from './types'
 
 export { TradeExecution } from './tradeExecution'
 
 export const getTradeQuotes = async (
   getTradeQuoteInput: GetTradeQuoteInput,
   enabledSwappers: SwapperName[],
-  deps: TradeQuoteDeps,
 ): Promise<QuoteResult[]> => {
   if (bnOrZero(getTradeQuoteInput.affiliateBps).lt(0)) return []
   if (getTradeQuoteInput.sellAmountIncludingProtocolFeesCryptoBaseUnit === '0') return []
@@ -22,7 +21,7 @@ export const getTradeQuotes = async (
       .filter(({ swapperName }) => enabledSwappers.includes(swapperName))
       .map(({ swapperName, swapper }) =>
         timeout<TradeQuote2[], SwapErrorRight>(
-          swapper.getTradeQuote(getTradeQuoteInput, deps),
+          swapper.getTradeQuote(getTradeQuoteInput),
           QUOTE_TIMEOUT_MS,
           QUOTE_TIMEOUT_ERROR,
         ).then(quote => ({
