+++ src/lib/swapper/swappers/OneInchSwapper/endpoints.ts
@@ -2,8 +2,6 @@ import type { EvmChainId } from '@shapeshiftoss/chain-adapters'
 import type { ETHSignTx } from '@shapeshiftoss/hdwallet-core'
 import type { Result } from '@sniptt/monads/build'
 import { v4 as uuid } from 'uuid'
-import { bnOrZero } from 'lib/bignumber/bignumber'
-import { fromBaseUnit } from 'lib/math'
 import type {
   GetEvmTradeQuoteInput,
   GetTradeQuoteInput,
@@ -12,7 +10,6 @@ import type {
   Swapper2Api,
   TradeQuote2,
 } from 'lib/swapper/api'
-import { getMinimumDonationUsdSellAmountByChainId } from 'lib/swapper/swappers/utils/getMinimumDonationUsdSellAmountByChainId'
 import { assertGetEvmChainAdapter, checkEvmSwapStatus } from 'lib/utils/evm'
 
 import { getTradeQuote } from './getTradeQuote/getTradeQuote'
@@ -23,31 +20,11 @@ const tradeQuoteMetadata: Map<string, { chainId: EvmChainId }> = new Map()
 export const oneInchApi: Swapper2Api = {
   getTradeQuote: async (
     input: GetTradeQuoteInput,
-    { sellAssetUsdRate }: { sellAssetUsdRate: string },
   ): Promise<Result<TradeQuote2[], SwapErrorRight>> => {
-    const {
-      sellAsset,
-      sellAmountIncludingProtocolFeesCryptoBaseUnit,
-      affiliateBps,
-      receiveAddress,
-    } = input
+    const { affiliateBps, receiveAddress } = input
 
-    const sellAmountBeforeFeesCryptoPrecision = fromBaseUnit(
-      sellAmountIncludingProtocolFeesCryptoBaseUnit,
-      sellAsset.precision,
-    )
-    const sellAmountBeforeFeesUsd = bnOrZero(sellAmountBeforeFeesCryptoPrecision).times(
-      sellAssetUsdRate,
-    )
-    // We use the sell amount so we don't have to make 2 network requests, as the receive amount requires a quote
-    const isDonationAmountBelowMinimum = sellAmountBeforeFeesUsd.lt(
-      getMinimumDonationUsdSellAmountByChainId(sellAsset.chainId),
-    )
+    const tradeQuoteResult = await getTradeQuote(input as GetEvmTradeQuoteInput)
 
-    const tradeQuoteResult = await getTradeQuote({
-      ...(input as GetEvmTradeQuoteInput),
-      affiliateBps: isDonationAmountBelowMinimum ? '0' : affiliateBps,
-    })
     return tradeQuoteResult.map(tradeQuote => {
       const id = uuid()
       const firstHop = tradeQuote.steps[0]
@@ -56,7 +33,7 @@ export const oneInchApi: Swapper2Api = {
         {
           id,
           receiveAddress,
-          affiliateBps: isDonationAmountBelowMinimum ? undefined : affiliateBps,
+          affiliateBps,
           ...tradeQuote,
         },
       ]
