+++ src/lib/swapper/swappers/ZrxSwapper/endpoints.ts
@@ -2,8 +2,6 @@ import type { EvmChainId } from '@shapeshiftoss/chain-adapters'
 import type { ETHSignTx } from '@shapeshiftoss/hdwallet-core'
 import type { Result } from '@sniptt/monads/build'
 import { v4 as uuid } from 'uuid'
-import { bnOrZero } from 'lib/bignumber/bignumber'
-import { fromBaseUnit } from 'lib/math'
 import type {
   GetEvmTradeQuoteInput,
   GetTradeQuoteInput,
@@ -12,7 +10,6 @@ import type {
   Swapper2Api,
   TradeQuote2,
 } from 'lib/swapper/api'
-import { getMinimumDonationUsdSellAmountByChainId } from 'lib/swapper/swappers/utils/getMinimumDonationUsdSellAmountByChainId'
 import { assertGetEvmChainAdapter, checkEvmSwapStatus } from 'lib/utils/evm'
 
 import { getZrxTradeQuote } from './getZrxTradeQuote/getZrxTradeQuote'
@@ -23,29 +20,10 @@ const tradeQuoteMetadata: Map<string, { chainId: EvmChainId }> = new Map()
 export const zrxApi: Swapper2Api = {
   getTradeQuote: async (
     input: GetTradeQuoteInput,
-    { sellAssetUsdRate }: { sellAssetUsdRate: string },
   ): Promise<Result<TradeQuote2[], SwapErrorRight>> => {
-    const {
-      sellAsset,
-      sellAmountIncludingProtocolFeesCryptoBaseUnit,
-      affiliateBps,
-      receiveAddress,
-    } = input
-    const sellAmountBeforeFeesCryptoPrecision = fromBaseUnit(
-      sellAmountIncludingProtocolFeesCryptoBaseUnit,
-      sellAsset.precision,
-    )
-    const sellAmountBeforeFeesUsd = bnOrZero(sellAmountBeforeFeesCryptoPrecision).times(
-      sellAssetUsdRate,
-    )
-    // We use the sell amount so we don't have to make 2 network requests, as the receive amount requires a quote
-    const isDonationAmountBelowMinimum = sellAmountBeforeFeesUsd.lt(
-      getMinimumDonationUsdSellAmountByChainId(sellAsset.chainId),
-    )
-    const tradeQuoteResult = await getZrxTradeQuote({
-      ...(input as GetEvmTradeQuoteInput),
-      affiliateBps: isDonationAmountBelowMinimum ? '0' : affiliateBps,
-    })
+    const { affiliateBps, receiveAddress } = input
+
+    const tradeQuoteResult = await getZrxTradeQuote(input as GetEvmTradeQuoteInput)
 
     return tradeQuoteResult.map(tradeQuote => {
       const id = uuid()
@@ -56,7 +34,7 @@ export const zrxApi: Swapper2Api = {
         {
           id,
           receiveAddress,
-          affiliateBps: isDonationAmountBelowMinimum ? undefined : affiliateBps,
+          affiliateBps,
           ...tradeQuote,
         },
       ]
